Allow limiting search time using fuel
To use Chalk in an IDE environment, we need to be able to prevent it from taking
too much time, even if that means getting a wrong or incomplete answer.
This is done by specifying an amount of 'fuel'; each iteration of the top-level
search loop takes one unit of fuel, and when it runs out, we return.